refactor: Rename utilities folder to internal
