+++ source/gallery.ts
@@ -1,12 +1,12 @@
 import type { Hitomi } from './hitomi';
 import { Image, Video } from './media';
 import { Language, Tag } from './tag';
-import { RESOURCE_DOMAIN, DEDICATED_TAG_PROPERTIES } from './utilities/constants';
+import { RESOURCE_DOMAIN, DEDICATED_TAG_PROPERTIES } from './internal/constants';
 import { SortType } from './enums';
-import { defineProperties, hashTerm, formatOneOfState, toString } from './utilities/functions';
-import { Base, IndexProvider } from './utilities/structures';
+import { defineProperties, hashTerm, formatOneOfState, toString } from './internal/functions';
+import { Base, IndexProvider } from './internal/structures';
 import { HitomiError } from './error';
-import type { URL, Node } from './utilities/types';
+import type { URL, Node } from './internal/types';
 
 /**
  * Title associated with a gallery.
