+++ source/index.ts
@@ -4,8 +4,8 @@ export type { Gallery, Title, TranslatedGallery, GalleryReference, GalleryManage
 export { Hitomi } from './hitomi';
 export type { Image, Video } from './media';
 export type { Language, Tag, TagManager } from './tag';
-export { HitomiError } from './utilities/structures';
-export { SortType, NameInitial, Extension, ThumbnailSize } from './utilities/constants';
+export { HitomiError } from './error';
+export { SortType, NameInitial, Extension, ThumbnailSize } from './enums';
 
 export const hitomi: Hitomi = new Hitomi();
 export default hitomi;
