+++ source/internal/functions.ts
