+++ source/internal/structures.ts
