+++ tests/utilities/constants.ts
@@ -1,6 +1,6 @@
 import { TextEncoder } from 'util';
 import type { Tag } from '../../source/tag';
-import { TAG_TYPES } from '../../source/utilities/constants';
+import { TAG_TYPES } from '../../source/internal/constants';
 
 export const PARTIAL_TAG_TYPES: Set<Tag['type']> = new Set(TAG_TYPES);
